pyproject.toml
[tool.poetry]
name = "home-work-30"
version = "0.1.0"
description = ""
authors = ["KHAGE <khazhanyants@mail.com>"]
readme = "README.md"
packages = [{include = "home_work_30"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
